fix:exception bugs

+++ src/main/java/com/khamidgaipov/api/giybat/uz/entity/ProfileRoleEntity.java
@@ -21,7 +21,7 @@ public class ProfileRoleEntity {
     private ProfileEntity profile;
 
     @Column(name = "profile_id")
-    private int profileId;
+    private Long profileId;
 
     @Column
     @Enumerated(EnumType.STRING)
